+++ core/libs/__init__.py
@@ -132,9 +132,12 @@ def create2DSystem(num: int) -> Tuple[System, Workplane, LineSegment2d]:
 
 def slvsProcess(
     vpoints: Tuple[VPoint],
-    constraints: Tuple[Tuple[int, str, str, float]]
+    constraints: Tuple[Tuple[int, int, float]]
 ) -> Tuple[List[Tuple[float, float]], int]:
     """Use element module to convert into solvespace expression."""
+    if not vpoints:
+        return [], 0
+    
     #Define VLinks here.
     vlinks = {}
     for i, vpoint in enumerate(vpoints):
@@ -146,15 +149,14 @@ def slvsProcess(
                 vlinks[linkName] = [i]
     
     #Limitation of Solvespacce kernel sys.
-    pointCount = 0
-    sliderCount = 0
+    point_count = 0
+    slider_count = 0
     for vpoint in vpoints:
-        pointCount += len(vpoint.c)
+        point_count += len(vpoint.c)
         if (vpoint.type == 1) or (vpoint.type == 2):
-            sliderCount += 1
-    constraintCount = len(constraints)
+            slider_count += 1
     
-    sys, wp1, h_line = create2DSystem(pointCount*2 + sliderCount*2 + constraintCount*2)
+    sys, wp1, h_line = create2DSystem(point_count * 2 + slider_count * 2 + len(constraints) * 2)
     
     solved_points = []
     
@@ -221,7 +223,7 @@ def slvsProcess(
                 else:
                     p_link_assist = solved_points[relate_n]
                 l_link = LineSegment2d(wp1, p_main, p_link_assist)
-                angle_base = vpoint.slopeAngle(relate_vp)
+                angle_base = vpoint.slope_angle(relate_vp)
                 if (angle_base == 0.) or (angle_base == 180.):
                     Constraint.parallel(wp1, l_link, l_slot)
                 else:
@@ -302,37 +304,36 @@ def slvsProcess(
             else:
                 ConnectTo(*connect_1)
     
-    for shaft, base_link, drive_link, angle in constraints:
+    for p0, p1, angle in constraints:
         """The constraints of drive shaft.
         
         Simulate the input variables to the mechanism.
         The 'base points' are shaft center.
         """
-        if vpoints[shaft].type != 0:
-            p_base = solved_points[shaft][0]
+        if vpoints[p0].type != 0:
+            p_base = solved_points[p0][0]
         else:
-            p_base = solved_points[shaft]
+            p_base = solved_points[p0]
         
         #Base link slope angle.
+        base_link = vpoints[p0].links[0]
         if base_link != 'ground':
             relate_base = vlinks[base_link]
-            newRelateOrder_base = relate_base.index(shaft)-1
-            angle -= vpoints[shaft].slopeAngle(vpoints[newRelateOrder_base])
+            newRelateOrder_base = relate_base.index(p0) - 1
+            angle -= vpoints[p0].slope_angle(vpoints[newRelateOrder_base])
             angle %= 360.
         
-        x = sys.add_param(round(vpoints[shaft].cx + 10.*cos(radians(angle)), 8))
-        y = sys.add_param(round(vpoints[shaft].cy + 10.*sin(radians(angle)), 8))
+        x = sys.add_param(round(vpoints[p0].cx + 10 * cos(radians(angle)), 8))
+        y = sys.add_param(round(vpoints[p0].cy + 10 * sin(radians(angle)), 8))
         p_hand = Point2d(wp1, x, y)
         Constraint.dragged(wp1, p_hand)
         #The virtual link that dragged by "hand".
         leader = LineSegment2d(wp1, p_base, p_hand)
         #Make another virtual link that should follow "hand".
-        relate_drive = vlinks[drive_link]
-        newRelateOrder_drive = relate_drive[relate_drive.index(shaft)-1]
-        if vpoints[newRelateOrder_drive].type != 0:
-            p_drive = solved_points[newRelateOrder_drive][0]
+        if vpoints[p1].type != 0:
+            p_drive = solved_points[p1][0]
         else:
-            p_drive = solved_points[newRelateOrder_drive]
+            p_drive = solved_points[p1]
         link = LineSegment2d(wp1, p_base, p_drive)
         Constraint.angle(wp1, .5, link, leader)